Switch to Williamson 2 for testing + fix numerical flux
The original numerical flux for the velocity was evidently too diffusive
-- the profile was getting smeared out and the system was obviously
losing lots of energy. Taking out a factor of the thickness from the
numerical flux seems to have fixed it in that the Williamson 2 setup is
steady, which it should be, but this makes no sense because the units
are all wrong.